+++ tests/input_files/series/pictorial/01.js
@@ -1,0 +1,14 @@
+{
+  id: 'some-id-goes-here',
+  index: 3,
+  legendIndex: 3,
+  name: 'Series Name Goes Here',
+  grouping: false,
+  groupPadding: 6,
+  maxPointWidth: 12,
+  minPointLength: 12,
+  pointPadding: 6,
+  pointRange: 24,
+  pointWidth: 12,
+  type: 'pictorial'
+}